Option to open modal in full screen as iframe (ES6)
Fixes #381
With this PR it is possible to config an url to be opened inside a modal window instead of a new window. Adding "target" = "iframe" will add a class to the url which will open the link as an iframe inside modal window.
Rewrote earlier PR according to es6