`groupSuggestions` option to active suggestion groups, with title same as popups
The `groupSuggestions` option is added for the Search control and replaces `layerNameAttribute` in controlling whether or not to group suggestions. The title for suggestion group is fetched the same way as the title for feature info popups, providing more consistency and flexibility (like the possibility to add suggestion group titles for single layer searches or searches without defined map layers.