feat: add feature styling
Fixes #1676
Adds the possibility to set the drag and drop control as well as a geojson source to style features based on the "style" attribute of the feature rather than the layer style.
Configured as `"styleByAttribute":true` as options for the drag and drop control or property of a geojson-layer.
False by default