Update guide menu item state based on modal removal
Observers if guide modal is removed from DOM and sets the proper state to guide menu item. If the menu item state is 'initial', it creates a modal and sets the state to 'active'. If the state is 'active', it closes the modal and sets the state back to 'initial'.