Add map name to favourite object before sending to server
This commit adds a line of code to include the map name in the favourite object before sending it to the server. The map name is obtained from the viewer and replaces the '.json' extension. This change ensures that the server has access to the map name when processing the favourite object.